Add Datamatic directory
